+++ qblox_drive_AS/Executions/m2_ZoomCS.py
@@ -1,0 +1,20 @@
+from qblox_drive_AS.support.ExpFrames import Zoom_CavitySearching
+from qblox_drive_AS.support.Path_Book import find_latest_QD_pkl_for_dr
+from qblox_drive_AS.support import Data_manager
+
+#TODO: TO test
+
+''' fill in '''
+Execution:bool = True
+DRandIP = {"dr":"dr2","last_ip":"10"}
+freq_range:dict = {"q0":[5.985e9, 5.995e9], "q1":[6.067e9, 6.077e9]}
+freq_pts:int = 100
+AVG:int = 100
+
+''' Don't Touch '''
+save_dir = Data_manager().build_packs_folder()
+EXP = Zoom_CavitySearching(QD_path=find_latest_QD_pkl_for_dr(DRandIP["dr"],DRandIP["last_ip"]),data_folder=save_dir)
+EXP.SetParameters(freq_range,freq_pts,AVG,Execution)
+EXP.WorkFlow()
+EXP.RunAnalysis()
+
